refactor: abstract repositories
