+++ app/src/main/java/woowacourse/shopping/data/DefaultProductRepository.kt
@@ -1,8 +1,9 @@
 package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.repository.ProductRepository
 
-class ProductRepository {
+class DefaultProductRepository : ProductRepository {
     private val products: List<Product> =
         listOf(
             Product(
@@ -22,7 +23,5 @@ class ProductRepository {
             ),
         )
 
-    fun getAllProducts(): List<Product> {
-        return products
-    }
+    override fun getAllProducts(): List<Product> = products
 }
