+++ app/src/main/java/woowacourse/shopping/repository/ProductRepository.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.repository
+
+import woowacourse.shopping.model.Product
+
+interface ProductRepository {
+    fun getAllProducts(): List<Product>
+}
