+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -2,16 +2,17 @@ package woowacourse.shopping.ui.cart
 
 import android.os.Bundle
 import android.widget.Toast
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
+import woowacourse.shopping.di.viewModelInject
 
 class CartActivity : AppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
 
-    private val viewModel by lazy {
-        ViewModelProvider(this)[CartViewModel::class.java]
+    private val viewModel: CartViewModel by viewModels {
+        viewModelInject<CartViewModel>()
     }
 
     private lateinit var dateFormatter: DateFormatter
